% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-unify_format.R
\name{unify_format}
\alias{unify_format}
\title{Unify the format of data}
\usage{
unify_format(raw_data)
}
\arguments{
\item{raw_data}{unformated data frame from excel}
}
\value{
formated data frame
}
\description{
Unify the format of data
}
